Custom JsonArrayRequest to retrieve Header Link
The default implementation of JsonArrayRequest is not
providing the Header. Extending it to be able to
retrieve the "Link" section, which is providing info
about the URL request to use for moving to first, prev,
next and last page. This is needed because I'm using
pagination when requesting the author list.