Implement custom ArrayAdapters for each Activity
The Activities are sharing the same Table, but each Activity
will insert a different type of row (with different types of
elements such as image, textView...).
In addition, when clicking on a given position of the Table,
each Adapter know how to retrieve the correct object (Author,
Comment, Position). This avoid the usage of unneeded global
list in each Activity to take track of the items shown in the table
at a given time.